TESTS/ after deploy bug fixes, changed image upload trait to not crop and resize test images, some visual changes on test proccess

+++ database/migrations/2019_09_26_112551_create_test_users_submited_table.php
@@ -20,7 +20,7 @@ class CreateTestUsersSubmitedTable extends Migration
             $table->dateTime('started_at')->nullable();
             $table->dateTime('submited_at')->nullable();
             $table->bigInteger('questions')->nullable();
-            $table->bigInteger('score')->nullable();
+            $table->float('score',10,2)->nullable();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
